% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article.R
\name{vub_article}
\alias{vub_article}
\title{Use article with the VUB theme}
\usage{
vub_article(
  subtitle,
  reportnr,
  faculty = "DIGI: Brussels Platform for Digital Humanities",
  floatbarrier = c(NA, "section", "subsection", "subsubsection"),
  codesize = c("footnotesize", "scriptsize", "tiny", "small", "normalsize"),
  keep_tex = FALSE,
  fig_crop = TRUE,
  includes = NULL,
  pandoc_args = NULL,
  ...
)
}
\arguments{
\item{subtitle}{An optional subtitle.}

\item{reportnr}{The report number. Defaults to the date and time of compilation.}

\item{faculty}{The faculty.}

\item{floatbarrier}{Should float barriers be placed?
Defaults to NA (only float barriers before starting a new chapter `#`).
Options are "section" (`##`), "subsection" (`###`) and
"subsubsection" (`####`).}

\item{codesize}{The fontsize of the code, relative to the fontsize of the text (= normal size).
Allowed values are "normalsize", "small", "footnotesize", "scriptsize" and "tiny". Defaults to "footnotesize".}

\item{keep_tex}{Keep the tex file. Defaults to FALSE.}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility
(if available) to pdf figures.}

\item{includes}{Passed on to \code{\link{includes_to_pandoc_args}}}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{...}{extra parameters: see details}
}
\description{
Use article with the VUB theme
}
\details{
Available extra parameters:
- `lof`: display a list of figures. Defaults to TRUE
- `lot`: display a list of tables. Defaults to TRUE
- `hyphenation`: the correct hyphenation for certain words.
- `cover`: an optional pdf file. The first two pages will be prepended to the
report.
}
\examples{
outformat <- vub_article()
}
