% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{vub_slides}
\alias{vub_slides}
\title{Use slides with the VUB theme}
\usage{
vub_slides(
  subtitle = "Provide a subtitle",
  location = "VUB - Brussels",
  institute,
  toc_name,
  slide_level = 2,
  fontsize,
  codesize = c("footnotesize", "scriptsize", "tiny", "small", "normalsize"),
  lang = "dutch",
  toc = FALSE,
  keep_tex = FALSE
)
}
\arguments{
\item{subtitle}{The subtitle.}

\item{location}{The date and place of the event.}

\item{institute}{The affiliation of the authors.}

\item{toc_name}{Name of the table of contents. Defaults to "Overzicht".}

\item{slide_level}{Indicate which heading level is used for the frame titles}

\item{fontsize}{The fontsite of the document. Defaults to 10pt.}

\item{codesize}{The fontsize of the code, relative to the fontsize of the text (= normal size).
Allowed values are "normalsize", "small", "footnotesize", "scriptsize" and "tiny". Defaults to "footnotesize".}

\item{lang}{The language of the document. Defaults to "dutch"}

\item{toc}{display a table of content after the title slide}

\item{keep_tex}{Keep the tex file. Defaults to FALSE.}

\item{...}{extra parameters}
}
\description{
Use slides with the VUB theme
}
\examples{
outformat <- vub_slides()
}
